got sitemaps working
